put familyresult into intermediate file
